rename comfy/logging.py to comfy/internal_logging.py
This avoids name collision (circular imports) for external custom nodes,
for which the comfy path is pushed into sys.path so Python's own logging module
is shadowed otherwise.

fixes: #15229
